Move cache database adapter
