+++ nautilus_trader/cache/adapter.py
